attempting chad-clumb suggestion to fix CRAN warning about gcc install error #3
