+++ ChangeLog
@@ -1,6 +1,7 @@
 Change Log:
 v0.10.1
-  - update get.vertex.attribute.active and list.vertex.attributes.active to be compatible with new network generics
+  - update get.vertex.attribute.active and list.vertex.attributes.active to be compatible with new network generics (thanks to chad-klumb!)
+  - combined spellfunctions.c into Rinit.c to avoid CRAN-flagged problems with -fno-common flag in gcc
 v0.10
   - update to test failing on cran after network 1.14 release
   - updates to network.edgecount.active, network.dyadcount.active and similar to be compatible
