+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: networkDynamic
 Version: 0.10.1
-Date: 2019-10-11
+Date: 2020-01-16
 Title: Dynamic Extensions for Network Objects
 Type: Package
 Depends: R (>= 3.0.0), network (>= 1.16.0)
